update event-west.yaml supporter tiers and dates
Removed Golden Bagel kick-in level. Updated the pre-reg open dates, as well as the badge price increase dates. I'm alright with using a dropdown instead of these big buttons for the tiers, but if that is too difficult could we grab the icons and html pages from this year's super mag? A link to where I can edit those would be much appreciated.